Rename callback files to more consistant style
