+++ netket/callbacks/early_stopping.py
