+++ netket/callbacks/invalid_loss_stopping.py
